Open tag panel by default
Left panel seems to be important tool for managing tags, so I think it should be open by default. 

It's especially important for the first time users. Opening panel is described in the initial guiding tasks, but let's make it as easy as possible to discover this important feature, even without getting through the guide.