Sorting Actionable from most urgent due dates
Related to https://github.com/getting-things-gnome/gtg/issues/859

Current sorting in Actionable view starts with no date tasks and most urgent dates go to the end, which means `today` might not be even visible. This simple patch reverses the sorting order of due date, so it starts with `today` tasks.