Remove task count for saved searches
Calculating a saved search's task count would involve running the search
and seeing how many tasks it returns. With several (and possibly
complex) searches this would result in a serious slowdown everytime a
task is modified, which is very often.

We can bring them back eventually if we find an efficient way to do
this.

ref #840