Write unit test for TaskView.get_title()
I started to look at #1077 and I thought it would be nice to be able to
unit test this type of thing.

A simple unit test for TaskView would be

* create a task
* create a view for it
* check that the title of the view is the title of the task

In the current state of affairs, this is difficult because the code that
puts the Task data into the view lives in TaskEditor. But we pass a Task
to the constructor of TaskView, so TaskView actually has all the data it
needs. So I moved the logic there. I also moved the `is_new()` logic
from the TaskEditor down to the Task.

I feel like it should be possible to eventually do this in the TaskView
constructor but for now, TaskEditor calls
`self.textview.set_text_from_task()` from the same place it used to run
that code.

The test triggers a segmentation fault when running on GitHub Actions,
so it's skipped there while I figure out why.