Add simple pre-commit hook
Relates to #237

This adds a simple configuration for pre-commit and runs it on GHA. It doesn't do very much; mostly removing trailing whitespaces.

My next step would be to add pylint or pyflakes to this, and tell it to ignore every warning that GTG currently breaks. This would:
* prevent us from breaking any more
* allow us to fix the warnings one at a time, when we can

I tested the GHA on a fork.

This commit doesn't contain the changes that fix the pre-commit check, so GHA will fail. To fix it, we'll just need to run `pre-commit run --all-files` and commit the result.