Update models for subtasks upon changes
The tid_to_subtask_model field keeps track of the ListStores
created for subtasks. The _remove_from_parent_model
and _append_to_model helper functions update these models
when a change happens. This functionality helps to keep
the UI up-to-date.

This fixes one of the problems listed in GitHub issue #1018