Apply simple Ruff fixes
Relates to #237

With one exception, this is the output of `ruff lint --fix`. The changes
look minor - it mostly removes unused imports.

Ruff reports a lot of other errors, including a fair amount of
"Undefined name" that indicate broken code paths. We can tackle those
separately, and decide whether we want to add Ruff to the pre-commit
hook.

This commit in itself is still a good quality of life improvement.

I also did some smoke testing to make sure nothing is obviously broken.