Use GNOME 46 in weekly flatpak
Relates to #708

This fixes the following error:
```
$ flatpak install org.gnome.GTG.Devel_20240530.flatpak
error: The application org.gnome.GTG.Devel/x86_64/master requires the runtime org.gnome.Platform/x86_64/master which was not found
```
by using the latest GNOME runtime, as opposed to the "master" one, which
doesn't seem to exist.

Tested by running the workflow on my fork, downloading the artifact from
GHA, installing and running the flatpak.