Improve version name
Update base version from 0.6.0 to 0.7-dev.

The idea is that when we're ready to release 0.7, we'll create a 0.7 
branch. We'll then update the version to "0.7" on that branch, and
"0.8-dev" on master.

Append the short commit hash in development mode (as before)

Append the date in flatpak mode, to help keep track of versions
when using the weekly flatpak. Tested the flatpak by running the GHA
workflow on my fork.