Update links to old wiki
This commit updates the references to wiki.gnome.org, except the ones in
the translations (I assume there is a process to mark translations as
outdated), and the following:

```
$ ag wiki.gnome.org --ignore po
docs/contributors/git workflow tips.md
31:Pro tip: using a graphical tool like [gitg](https://wiki.gnome.org/Apps/Gitg/), which not only lets you see visually the relationship between the various branches, but also lets you select which lines of which files you want to "stage" to commit, making it INCREDIBLY easier, and letting you make much cleaner, smaller, "atomic" commits. Check out the [Meld](https://meldmerge.org/) tool, too.

docs/contributors/profiling GTG for performance.md
63:![Generated image](https://wiki.gnome.org/Apps/GTG/development?action=AttachFile&do=get&target=profile.png)

docs/source/releases/v0.3.1.rst
41:.. _`his Google Summer of Code 2013 project`: https://wiki.gnome.org/Outreach/SummerOfCode/2013/Projects/ParinPorecha_GTGOnline
50:.. _`ported GTG to python3 and gtk3`: https://wiki.gnome.org/Outreach/SummerOfCode/2013/Projects/XuanHu_PortingGTG
107:.. _`collaborative task management`: https://wiki.gnome.org/Outreach/SummerOfCode/2012/Projects/IzidorMatusov_CollaborativeGTG
```

gitg still marks wiki.gnome.org as their website, so I didn't change the
link. Updating it to the GitLab project would also be a viable option.

The image was probably lost to time and would need to be regenerated.

The 0.3.1 release references some GSoC projects that I haven't chased
new links for.