Avoid deleting children multiple times while purging old tasks
Use BaseStore.batch_remove to avoid deleting the same closed task
multiple times in Application.purge_old_tasks. (E.g., the task is
first deleted as the child of another closed task, and then a
second delete is attempted because the task is also present in
'to_remove').

This fixes GitHub issue #1247.